Updated Moderate responders in CleanMetadata
A couple of changes…

1. the function is now called tnfa_response() so that I can add other
data sets in that file for you

2. The function now take an argument called Moderates which by default
is set to False. If Moderates is false it will not include any
moderates. If it is set to 1, it will include them and group them with
the responders. If it is set to 2, it will create a new group where
medium response = 2 (i know it’s non intuitive but it was a lot easier
that way.)